+++ CheckInSKP/Infrastructure/Repositories/StaffRepository.cs
@@ -21,20 +21,15 @@ namespace CheckInSKP.Infrastructure.Repositories
 
         public async Task AddAsync(Staff staff)
         {
-            if (staff == null)
-            {
-                throw new ArgumentNullException(nameof(staff));
-            }
-
-            StaffEntity entity = _staffMapper.MapToEntity(staff);
-            await _context.Staffs.AddAsync(entity);
+            var entity = _staffMapper.MapToEntity(staff);
+            _ = await _context.Staffs.AddAsync(entity);
         }
 
-        public async Task<Staff> GetByIdAsync(int id)
+        public async Task<Staff?> GetByIdAsync(int id)
         {
-            StaffEntity entity = await _context.Set<StaffEntity>()
+            var entity = await _context.Set<StaffEntity>()
                                                .Include(e => e.TimeLogs)
-                                               .FirstOrDefaultAsync(e => e.Id == id) ?? throw new EntityNotFoundException($"Staff with id {id} not found.");
+                                               .FirstOrDefaultAsync(e => e.Id == id);
 
             return _staffMapper.MapToDomain(entity);
         }
@@ -42,65 +37,64 @@ namespace CheckInSKP.Infrastructure.Repositories
 
         public async Task UpdateAsync(Staff staff)
         {
-            StaffEntity entity = await _context.Set<StaffEntity>().FindAsync(staff.Id) ?? throw new EntityNotFoundException($"Staff with id {staff.Id} not found.");
+            var entity = await _context.Set<StaffEntity>().FindAsync(staff.Id);
 
-            entity = _staffMapper.MapToEntity(staff);
-            _context.Entry(entity).State = EntityState.Modified;
+            if(entity != null)
+            {
+                entity.PhoneNumber = staff.PhoneNumber;
+                entity.CardNumber = staff.CardNumber;
+                entity.PhoneNotification = staff.PhoneNotification;
+                entity.Preoccupied = staff.Preoccupied;
+                entity.MeetingTime = staff.MeetingTime;
+                entity.UserId = staff.UserId;
+
+                // TODO: TimeLog fix
+            }
         }
 
         public async Task RemoveAsync(int id)
         {
-            StaffEntity entity = await _context.Set<StaffEntity>().FindAsync(id) ?? throw new EntityNotFoundException($"Staff with id {id} not found.");
-
-            _context.Set<StaffEntity>().Remove(entity);
+            var entity = await _context.Set<StaffEntity>().FindAsync(id);
+            if(entity != null)
+            {
+                _context.Set<StaffEntity>().Remove(entity);
+            }
         }
 
         public async Task<bool> ExistsAsync(int id)
         {
-            return await _context.Set<StaffEntity>().FindAsync(id) != null ? true : false;
+            return await _context.Set<StaffEntity>().FindAsync(id) != null;
         }
 
-        public async Task<IEnumerable<Staff>> GetAllAsync()
+        public async Task<IEnumerable<Staff?>> GetAllAsync()
         {
-            List<StaffEntity> entities = await _context.Set<StaffEntity>()
+            var entities = await _context.Set<StaffEntity>()
                                                        .Include(e => e.TimeLogs)
-                                                       .ToListAsync() ?? throw new EntityNotFoundException("No staff found.");
-            return entities.Select(e => _staffMapper.MapToDomain(e));
+                                                       .ToListAsync();
+            return entities.Select(_staffMapper.MapToDomain);
+        }
+
+        public async Task<IEnumerable<Staff?>> GetAllWithPaginationAsync(int page, int pageSize)
+        {
+            var entities = await _context.Set<StaffEntity>().Include(e => e.TimeLogs).Skip((page - 1) * pageSize).Take(pageSize).ToListAsync();
+            return entities.Select(_staffMapper.MapToDomain);
         }
 
         public async Task AddRangeAsync(IEnumerable<Staff> staffs)
         {
-            foreach (var staff in staffs)
-            {
-                if (staff == null)
-                {
-                    throw new ArgumentNullException(nameof(staff));
-                }
-            }
-
-            List<StaffEntity> entities = staffs.Select(_staffMapper.MapToEntity).ToList();
+            var entities = staffs.Select(_staffMapper.MapToEntity).ToList();
             await _context.Staffs.AddRangeAsync(entities);
         }
 
         public async Task RemoveRangeAsync(IEnumerable<int> staffIds)
         {
-            List<StaffEntity> entities = await _context.Staffs.Where(u => staffIds.Contains(u.Id)).ToListAsync() ?? throw new EntityNotFoundException("No staff found.");
-            _context.Staffs.RemoveRange(entities);
+            var entities = await _context.Staffs.Where(u => staffIds.Contains(u.Id)).ToListAsync();
+            if(entities != null)
+            {
+                _context.Staffs.RemoveRange(entities);
+            }
         }
 
-        public IQueryable<Staff> Query()
-        {
-            return _context.Set<StaffEntity>().Select(e => _staffMapper.MapToDomain(e)) ?? throw new EntityNotFoundException("No staff found.");
-        }
-
-        public async Task<IEnumerable<Staff>> GetAllWithPaginationAsync(int page, int pageSize)
-        {
-            List<StaffEntity> entities = await _context.Set<StaffEntity>()
-                                                       .Include(e => e.TimeLogs)
-                                                       .Skip((page - 1) * pageSize)
-                                                       .Take(pageSize)
-                                                       .ToListAsync() ?? throw new EntityNotFoundException("No staff found.");
-            return entities.Select(e => _staffMapper.MapToDomain(e));
-        }
+        public IQueryable<Staff?> Query() => _context.Set<StaffEntity>().Select(e => _staffMapper.MapToDomain(e));
     }
 }
