+++ CheckInSKP/Infrastructure/Repositories/UserRepository.cs
@@ -6,6 +6,7 @@ using CheckInSKP.Infrastructure.Entities;
 using CheckInSKP.Infrastructure.Exceptions;
 using CheckInSKP.Infrastructure.Mappings;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Cryptography.X509Certificates;
 
 namespace CheckInSKP.Infrastructure.Repositories
 {
@@ -22,20 +23,15 @@ namespace CheckInSKP.Infrastructure.Repositories
 
         public async Task AddAsync(User user)
         {
-            if (user == null)
-            {
-                throw new ArgumentNullException(nameof(user));
-            }
-
-            UserEntity entity = _userMapper.MapToEntity(user);
-            await _context.Users.AddAsync(entity);
+            var entity = _userMapper.MapToEntity(user);
+            _ = await _context.Users.AddAsync(entity);
         }
 
-        public async Task<User> GetByIdAsync(int id)
+        public async Task<User?> GetByIdAsync(int id)
         {
-            UserEntity entity = await _context.Set<UserEntity>()
+            var entity = await _context.Set<UserEntity>()
                                                .Include(e => e.Tokens)
-                                               .FirstOrDefaultAsync(e => e.Id == id) ?? throw new EntityNotFoundException($"User with id {id} not found.");
+                                               .FirstOrDefaultAsync(e => e.Id == id);
 
             return _userMapper.MapToDomain(entity);
         }
@@ -43,64 +39,66 @@ namespace CheckInSKP.Infrastructure.Repositories
 
         public async Task UpdateAsync(User user)
         {
-            UserEntity entity = await _context.Set<UserEntity>().FindAsync(user.Id) ?? throw new EntityNotFoundException($"User with id {user.Id} not found.");
+            var entity = await _context.Set<UserEntity>().FindAsync(user.Id);
+            if (entity != null)
+            {
+                entity.Name = user.Name;
+                entity.Username = user.Username;
+                entity.PasswordHash = user.PasswordHash;
+                entity.RoleId = user.RoleId;
 
-            entity = _userMapper.MapToEntity(user);
-            _context.Entry(entity).State = EntityState.Modified;
+                // TODO: Fix tokens
+            }
         }
 
         public async Task RemoveAsync(int id)
         {
-            UserEntity entity = await _context.Set<UserEntity>().FindAsync(id) ?? throw new EntityNotFoundException($"User with id {id} not found.");
-
-            _context.Set<UserEntity>().Remove(entity);
+            var entity = await _context.Set<UserEntity>().FindAsync(id);
+            if(entity != null)
+            {
+                _context.Set<UserEntity>().Remove(entity);
+            }
         }
 
         public async Task<bool> ExistsAsync(int id)
         {
-            return await _context.Set<UserEntity>().FindAsync(id) != null ? true : false;
+            return await _context.Set<UserEntity>().FindAsync(id) != null;
         }
 
-        public async Task<IEnumerable<User>> GetAllAsync()
+        public async Task<IEnumerable<User?>> GetAllAsync()
         {
-            List<UserEntity> entities = await _context.Set<UserEntity>().Include(e => e.Tokens).ToListAsync() ?? throw new EntityNotFoundException("No users found.");
-            return entities.Select(e => _userMapper.MapToDomain(e));
+            var entities = await _context.Set<UserEntity>().Include(e => e.Tokens).ToListAsync();
+            return entities.Select(_userMapper.MapToDomain);
+        }
+
+        public async Task<IEnumerable<User?>> GetAllWithPaginationAsync(int page, int pageSize)
+        {
+            var entities = await _context.Set<UserEntity>().Include(e => e.Tokens).Skip((page - 1) * pageSize).Take(pageSize).ToListAsync();
+            return entities.Select(_userMapper.MapToDomain);
         }
 
         public async Task AddRangeAsync(IEnumerable<User> users)
         {
-            foreach (var user in users)
-            {
-                if (user == null)
-                {
-                    throw new ArgumentNullException(nameof(user));
-                }
-            }
-
-            List<UserEntity> entities = users.Select(_userMapper.MapToEntity).ToList();
+            var entities = users.Select(_userMapper.MapToEntity).ToList();
             await _context.Users.AddRangeAsync(entities);
         }
 
         public async Task RemoveRangeAsync(IEnumerable<int> userIds)
         {
-            List<UserEntity> entities = await _context.Users.Where(u => userIds.Contains(u.Id)).ToListAsync() ?? throw new EntityNotFoundException("No users found.");
-            _context.Users.RemoveRange(entities);
+            var entities = await _context.Users.Where(u => userIds.Contains(u.Id)).ToListAsync();
+            if(entities != null)
+            {
+                _context.Users.RemoveRange(entities);
+            }
         }
 
-        public IQueryable<User> Query()
-        {
-            return _context.Set<UserEntity>().Select(e => _userMapper.MapToDomain(e)) ?? throw new EntityNotFoundException("No users found.");
-        }
+        public IQueryable<User?> Query() => _context.Set<UserEntity>().Select(e => _userMapper.MapToDomain(e));
 
-        public async Task<IEnumerable<User>> GetAllWithPaginationAsync(int page, int pageSize)
-        {
-            List<UserEntity> entities = await _context.Set<UserEntity>().Include(e => e.Tokens).Skip((page - 1) * pageSize).Take(pageSize).ToListAsync() ?? throw new EntityNotFoundException("No users found.");
-            return entities.Select(e => _userMapper.MapToDomain(e));
-        }
 
-        public Task<User> GetByUsernameAsync(string username)
+        public async Task<User?> GetByUsernameAsync(string username)
         {
-            return _context.Set<UserEntity>().Include(e => e.Tokens).FirstOrDefaultAsync(e => e.Username == username).ContinueWith(t => _userMapper.MapToDomain(t.Result)) ?? throw new EntityNotFoundException($"User with username {username} not found.");
+            var entity = await _context.Set<UserEntity>().FirstOrDefaultAsync(e => e.Username == username);
+            return _userMapper.MapToDomain(entity);
         }
     }
 }
